series

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -2,9 +2,9 @@
 
 int main(void)
 {
-int a = 0;
-int s = 1;
-int d;
+long int a = 0;
+long int s = 1;
+long int d;
 int f;
 for (f = 1; f < 98; f++)
 	{
